+++ CHANGELOG.md
@@ -5,7 +5,7 @@ this file.
 ## [UNRELEASED]
 - Added multi-channel processing support.
 - Added stereo/mid-side "balance" controls.
-- Added CLAP plugin format support.
+- Added CLAP plugin format support (with Parameter modulation).
 - Added option to enable/disable OpenGL on Windows/Linux systems with sufficient OpenGL support.
 - Improved presets system, and added new built-in presets from Carter and Cool WAV.
 - Fixed V1 mode creating high-pitched tone with linear phase oversampling.